requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
Django==4.1.7
django-admin==2.0.2
django-cors-headers==3.14.0
django-deep-serializer==0.1.3
django-excel-response2==3.0.5
django-redis==5.2.0
django-six==1.0.5
djangorestframework==3.14.0
excel-base==1.0.3
isoweek==1.3.3
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
redis==4.5.1
screen==1.0.1
six==1.16.0
sqlparse==0.4.3
TimeConvert==3.0.13
tzdata==2022.7
tzlocal==4.2
xlwt==1.3.0
